Rename rule to `no-unrestored-sinon-before-expect`
This makes more sense since it deals with stubs and spies and not just
stubs
